Separate tests for each lint
